Split imix exec function out.
